Create separate background proxy
